// src/gen/types.ts
export type ItemReferenceParam = {
  type?: ItemReferenceParamTypeEnumKey | null
  /**
   * @description The ID of the item to reference.
   * @type string
   */
  id: string
}
// ...
export type ReasoningItemParam = {
  id?: string | null
  /**
   * @description The item type. Always `reasoning`.
   * @default "reasoning"
   * @type string
   */
  type: ReasoningItemParamTypeEnumKey
  /**
   * @description Reasoning summary content associated with this item.
   * @type array
   */
  summary: Array<ReasoningSummaryContentParam>
  content?: null
  encrypted_content?: string | null
}
// ...
export type AssistantMessageItemParam = {
  id?: string | null
  /**
   * @description The item type. Always `message`.
   * @default "message"
   * @type string
   */
  type: AssistantMessageItemParamTypeEnumKey
  /**
   * @description The role of the message author. Always `assistant`.
   * @default "assistant"
   * @type string
   */
  role: AssistantMessageItemParamRoleEnumKey
  /**
   * @description The message content, as an array of content parts.
   */
  content:
    | Array<
        | (OutputTextContentParam & {
            type: 'output_text'
          })
        | (RefusalContentParam & {
            type: 'refusal'
          })
      >
    | string
  status?: string | null
}
// ...
export type FunctionCallItemParam = {
  id?: string | null
  /**
   * @description The unique ID of the function tool call generated by the model.
   * @minLength 1
   * @maxLength 64
   * @type string
   */
  call_id: string
  /**
   * @description The item type. Always `function_call`.
   * @default "function_call"
   * @type string
   */
  type: FunctionCallItemParamTypeEnumKey
  /**
   * @description The name of the function to call.
   * @minLength 1
   * @maxLength 64
   * @pattern ^[a-zA-Z0-9_-]+$
   * @type string
   */
  name: string
  /**
   * @description The function arguments as a JSON string.
   * @type string
   */
  arguments: string
  status?: (FunctionCallStatus & any) | null
}
// ...
export type FunctionCallOutputItemParam = {
  id?: string | null
  /**
   * @description The unique ID of the function tool call generated by the model.
   * @minLength 1
   * @maxLength 64
   * @type string
   */
  call_id: string
  /**
   * @description The type of the function tool call output. Always `function_call_output`.
   * @default "function_call_output"
   * @type string
   */
  type: FunctionCallOutputItemParamTypeEnumKey
  /**
   * @description Text, image, or file output of the function tool call.
   */
  output:
    | string
    | Array<
        | (InputTextContentParam & {
            type: 'input_text'
          })
        | (InputImageContentParamAutoParam & {
            type: 'input_image'
          })
        | (InputFileContentParam & {
            type: 'input_file'
          })
        | (InputVideoContent & {
            type: 'input_video'
          })
      >
  status?: (FunctionCallStatus & any) | null
}

export type ItemParam =
  | (ItemReferenceParam & {
      type: 'Item reference'
    })
  | (ReasoningItemParam & {
      type: 'reasoning'
    })
  | (AssistantMessageItemParam & {
      type: 'message'
    })
  | (FunctionCallItemParam & {
      type: 'function_call'
    })
  | (FunctionCallOutputItemParam & {
      type: 'function_call_output'
    })

export const includeEnumEnum = {
  'reasoning.encrypted_content': 'reasoning.encrypted_content',
  'message.output_text.logprobs': 'message.output_text.logprobs',
} as const

export type IncludeEnumEnumKey =
  (typeof includeEnumEnum)[keyof typeof includeEnumEnum]

export type IncludeEnum = IncludeEnumEnumKey

export type EmptyModelParam = object

export const functionToolParamTypeEnum = {
  function: 'function',
} as const

export type FunctionToolParamTypeEnumKey =
  (typeof functionToolParamTypeEnum)[keyof typeof functionToolParamTypeEnum]

export type FunctionToolParam = {
  /**
   * @minLength 1
   * @maxLength 64
   * @pattern ^[a-zA-Z0-9_-]+$
   * @type string
   */
  name: string
  description?: string | null
  parameters?: EmptyModelParam | null
  /**
   * @type boolean | undefined
   */
  strict?: boolean
  /**
   * @default "function"
   * @type string
   */
  type: FunctionToolParamTypeEnumKey
}

export type ResponsesToolParam = FunctionToolParam & {
  type: 'function'
}
// ...
export type SpecificToolChoiceParam = SpecificFunctionParam

export const toolChoiceValueEnumEnum = {
  none: 'none',
  auto: 'auto',
  required: 'required',
} as const

export type ToolChoiceValueEnumEnumKey =
  (typeof toolChoiceValueEnumEnum)[keyof typeof toolChoiceValueEnumEnum]

export type ToolChoiceValueEnum = ToolChoiceValueEnumEnumKey

export const allowedToolsParamTypeEnum = {
  allowed_tools: 'allowed_tools',
} as const

export type AllowedToolsParamTypeEnumKey =
  (typeof allowedToolsParamTypeEnum)[keyof typeof allowedToolsParamTypeEnum]

export type AllowedToolsParam = {
  /**
   * @description The tool choice type. Always `allowed_tools`.
   * @default "allowed_tools"
   * @type string
   */
  type: AllowedToolsParamTypeEnumKey
  /**
   * @description The list of tools that are permitted for this request.
   * @type array
   */
  tools: Array<SpecificToolChoiceParam>
  mode?: ToolChoiceValueEnum & any
}

/**
 * @description Controls which tool the model should use, if any.
 */
export type ToolChoiceParam =
  | SpecificToolChoiceParam
  | ToolChoiceValueEnum
  | AllowedToolsParam

/**
 * @description Set of 16 key-value pairs that can be attached to an object. This can be         useful for storing additional information about the object in a structured         format, and querying for objects via API or the dashboard.\n        Keys are strings with a maximum length of 64 characters. Values are strings         with a maximum length of 512 characters.
 */
export type MetadataParam = {
  [key: string]: string
}

export const verbosityEnumEnum = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const

export type VerbosityEnumEnumKey =
  (typeof verbosityEnumEnum)[keyof typeof verbosityEnumEnum]

export type VerbosityEnum = VerbosityEnumEnumKey
// ...
export type TextFormatParam = TextResponseFormat | JsonSchemaResponseFormatParam

export type TextParam = {
  /**
   * @description The format configuration for text output.
   */
  format?: TextFormatParam | null
  verbosity?: VerbosityEnum & any
}

/**
 * @description Options that control streamed response behavior.
 */
export type StreamOptionsParam = {
  /**
   * @description Whether to obfuscate sensitive information in streamed output. Defaults to `true`.
   * @type boolean | undefined
   */
  include_obfuscation?: boolean
}

export const reasoningEffortEnumEnum = {
  none: 'none',
  low: 'low',
  medium: 'medium',
  high: 'high',
  xhigh: 'xhigh',
} as const

export type ReasoningEffortEnumEnumKey =
  (typeof reasoningEffortEnumEnum)[keyof typeof reasoningEffortEnumEnum]

export type ReasoningEffortEnum = ReasoningEffortEnumEnumKey

export const reasoningSummaryEnumEnum = {
  concise: 'concise',
  detailed: 'detailed',
  auto: 'auto',
} as const

export type ReasoningSummaryEnumEnumKey =
  (typeof reasoningSummaryEnumEnum)[keyof typeof reasoningSummaryEnumEnum]

export type ReasoningSummaryEnum = ReasoningSummaryEnumEnumKey

/**
 * @description **gpt-5 and o-series models only** Configuration options for [reasoning models](https://platform.openai.com/docs/guides/reasoning).
 */
export type ReasoningParam = {
  effort?: ReasoningEffortEnum | null
  summary?: (ReasoningSummaryEnum & any) | null
}

export const truncationEnumEnum = {
  auto: 'auto',
  disabled: 'disabled',
} as const

export type TruncationEnumEnumKey =
  (typeof truncationEnumEnum)[keyof typeof truncationEnumEnum]

export type TruncationEnum = TruncationEnumEnumKey

export const serviceTierEnumEnum = {
  auto: 'auto',
  default: 'default',
  flex: 'flex',
  priority: 'priority',
} as const

export type ServiceTierEnumEnumKey =
  (typeof serviceTierEnumEnum)[keyof typeof serviceTierEnumEnum]

export type ServiceTierEnum = ServiceTierEnumEnumKey

export type CreateResponseBody = {
  model?: string | null
  input?: (string | Array<ItemParam>) | null
  previous_response_id?: string | null
  /**
   * @type array | undefined
   */
  include?: Array<IncludeEnum>
  tools?: Array<ResponsesToolParam> | null
  tool_choice?: (ToolChoiceParam & any) | null
  metadata?: (MetadataParam & any) | null
  text?: (TextParam & any) | null
  temperature?: number | null
  top_p?: number | null
  presence_penalty?: number | null
  frequency_penalty?: number | null
  parallel_tool_calls?: boolean | null
  /**
   * @description Whether to stream response events as server-sent events.
   * @type boolean | undefined
   */
  stream?: boolean
  stream_options?: (StreamOptionsParam & any) | null
  /**
   * @description Whether to run the request in the background and return immediately.
   * @type boolean | undefined
   */
  background?: boolean
  max_output_tokens?: number | null
  max_tool_calls?: number | null
  reasoning?: (ReasoningParam & any) | null
  safety_identifier?: string | null
  prompt_cache_key?: string | null
  truncation?: TruncationEnum & any
  instructions?: string | null
  /**
   * @description Whether to store the response so it can be retrieved later.
   * @type boolean | undefined
   */
  store?: boolean
  service_tier?: ServiceTierEnum & any
  top_logprobs?: number | null
}

/**
 * @description Details about why the response was incomplete.
 */
export type IncompleteDetails = {
  /**
   * @description The reason the response could not be completed.
   * @type string
   */
  reason: string
}
// ...
export type ItemField =
  | (Message & {
      type: 'message'
    })
  | (FunctionCall & {
      type: 'function_call'
    })
  | (FunctionCallOutput & {
      type: 'function_call_output'
    })
  | (ReasoningBody & {
      type: 'reasoning'
    })

/**
 * @description An error that occurred while generating the response.
 */
export type Error = {
  /**
   * @description A machine-readable error code that was returned.
   * @type string
   */
  code: string
  /**
   * @description A human-readable description of the error that was returned.
   * @type string
   */
  message: string
}
// ...
export type Tool = FunctionTool & {
  type: 'function'
}

export const functionToolChoiceTypeEnum = {
  function: 'function',
} as const

export type FunctionToolChoiceTypeEnumKey =
  (typeof functionToolChoiceTypeEnum)[keyof typeof functionToolChoiceTypeEnum]

export type FunctionToolChoice = {
  /**
   * @default "function"
   * @type string
   */
  type: FunctionToolChoiceTypeEnumKey
  /**
   * @type string | undefined
   */
  name?: string
}

export const allowedToolChoiceTypeEnum = {
  allowed_tools: 'allowed_tools',
} as const

export type AllowedToolChoiceTypeEnumKey =
  (typeof allowedToolChoiceTypeEnum)[keyof typeof allowedToolChoiceTypeEnum]

export type AllowedToolChoice = {
  /**
   * @default "allowed_tools"
   * @type string
   */
  type: AllowedToolChoiceTypeEnumKey
  /**
   * @type array
   */
  tools: Array<FunctionToolChoice>
  /**
   * @type string
   */
  mode: ToolChoiceValueEnum
}
// ...
export type TextField = {
  format:
    | TextResponseFormat
    | JsonObjectResponseFormat
    | JsonSchemaResponseFormat
  /**
   * @type string | undefined
   */
  verbosity?: VerbosityEnum
}

/**
 * @description Reasoning configuration and metadata that were used for the response.
 */
export type Reasoning = {
  effort: ReasoningEffortEnum | null
  summary: (ReasoningSummaryEnum & any) | null
}
// ...
export type Usage = {
  /**
   * @description The number of input tokens that were used to generate the response.
   * @type integer
   */
  input_tokens: number
  /**
   * @description The number of output tokens that were generated by the model.
   * @type integer
   */
  output_tokens: number
  /**
   * @description The total number of tokens that were used.
   * @type integer
   */
  total_tokens: number
  input_tokens_details: InputTokensDetails & any
  output_tokens_details: OutputTokensDetails & any
}

export const responseResourceObjectEnum = {
  response: 'response',
} as const

export type ResponseResourceObjectEnumKey =
  (typeof responseResourceObjectEnum)[keyof typeof responseResourceObjectEnum]

/**
 * @description The complete response object that was returned by the Responses API.
 * @example [object Object]
 */
export type ResponseResource = {
  /**
   * @description The unique ID of the response that was created.
   * @type string
   */
  id: string
  /**
   * @description The object type, which was always `response`.
   * @default "response"
   * @type string
   */
  object: ResponseResourceObjectEnumKey
  /**
   * @description The Unix timestamp (in seconds) for when the response was created.
   * @type integer
   */
  created_at: number
  completed_at: number | null
  /**
   * @description The status that was set for the response.
   * @type string
   */
  status: string
  incomplete_details: (IncompleteDetails & any) | null
  /**
   * @description The model that generated this response.
   * @type string
   */
  model: string
  previous_response_id: string | null
  instructions: string | null
  /**
   * @description The output items that were generated by the model.
   * @type array
   */
  output: Array<ItemField>
  error: (Error & any) | null
  /**
   * @description The tools that were available to the model during response generation.
   * @type array
   */
  tools: Array<Tool>
  tool_choice: FunctionToolChoice | ToolChoiceValueEnum | AllowedToolChoice
  truncation: TruncationEnum & any
  /**
   * @description Whether the model was allowed to call multiple tools in parallel.
   * @type boolean
   */
  parallel_tool_calls: boolean
  text: TextField & any
  /**
   * @description The nucleus sampling parameter that was used for this response.
   * @type number
   */
  top_p: number
  /**
   * @description The presence penalty that was used to penalize new tokens based on whether they appear in the text so far.
   * @type number
   */
  presence_penalty: number
  /**
   * @description The frequency penalty that was used to penalize new tokens based on their frequency in the text so far.
   * @type number
   */
  frequency_penalty: number
  /**
   * @description The number of most likely tokens that were returned at each position, along with their log probabilities.
   * @type integer
   */
  top_logprobs: number
  /**
   * @description The sampling temperature that was used for this response.
   * @type number
   */
  temperature: number
  reasoning: (Reasoning & any) | null
  usage: (Usage & any) | null
  max_output_tokens: number | null
  max_tool_calls: number | null
  /**
   * @description Whether this response was stored so it can be retrieved later.
   * @type boolean
   */
  store: boolean
  /**
   * @description Whether this request was run in the background.
   * @type boolean
   */
  background: boolean
  /**
   * @description The service tier that was used for this response.
   * @type string
   */
  service_tier: string
  /**
   * @description Developer-defined metadata that was associated with the response.
   */
  metadata: any
  safety_identifier: string | null
  prompt_cache_key: string | null
}